+++ src/main/java/com/example/momobe/user/domain/CustomUserRepository.java
@@ -1,7 +1,5 @@
 package com.example.momobe.user.domain;
 
-import com.example.momobe.user.domain.User;
-
 import java.util.Optional;
 
 public interface CustomUserRepository {
