+++ src/main/java/com/example/momobe/user/domain/Nickname.java
@@ -4,11 +4,14 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Access;
+import javax.persistence.AccessType;
 import javax.persistence.Embeddable;
 
 import static lombok.AccessLevel.*;
 
 @Embeddable
+@Access(AccessType.FIELD)
 @NoArgsConstructor(access = PROTECTED)
 public class Nickname {
     @Getter
