+++ src/test/java/com/example/momobe/user/domain/PasswordTest.java
@@ -1,6 +1,5 @@
 package com.example.momobe.user.domain;
 
-import com.example.momobe.common.enums.TestConstants;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -9,7 +8,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import static com.example.momobe.common.enums.TestConstants.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PasswordTest {
     PasswordEncoder passwordEncoder;
@@ -23,10 +21,10 @@ class PasswordTest {
     @DisplayName("패스워드가 일치하지 않을 경우 false 반환")
     public void match1() {
         //given
-        Password password = new Password(PASSWORD1, passwordEncoder);
+        Password password = new Password(PASSWORD1);
 
         //when
-        boolean result = password.match(PASSWORD2, passwordEncoder);
+        boolean result = password.match(PASSWORD2);
 
         //then
         Assertions.assertThat(result).isFalse();
@@ -36,10 +34,10 @@ class PasswordTest {
     @DisplayName("패스워드가 일치할 경우 true 반환")
     public void match2() {
         //given
-        Password password = new Password(PASSWORD1, passwordEncoder);
+        Password password = new Password(PASSWORD1);
 
         //when
-        boolean result = password.match(PASSWORD1, passwordEncoder);
+        boolean result = password.match(PASSWORD1);
 
         //then
         Assertions.assertThat(result).isTrue();
